+++ src/ClientBase.ts
@@ -6,6 +6,7 @@ import {
   HttpRequestHandler,
   DomainUtils,
   AccessKey,
+  ApiException as ApiExceptionCore
 } from '@laserfiche/lf-api-client-core';
 class ClientBase {}
 export interface IRepositoryApiClient {
@@ -1720,4 +1721,73 @@ export class LinkDefinitionsClient extends generated.LinkDefinitionsClient imple
       maxPageSize
     );
   }
+}
+
+export class CreateEntryResult extends generated.CreateEntryResult {
+  /** @internal */
+  getSummary(): string {
+    let messages = [];
+    const entryId: number = this.operations?.entryCreate?.entryId ?? 0;
+    if (entryId !== 0) {
+      messages.push(`entryId = ${entryId}`);
+    }
+
+    function getErrorMessages(errors: generated.APIServerException[] | undefined): string {
+      if (errors == null) {
+        return "";
+      }
+      
+      return errors.map(item => item.message).join(" ");
+    }
+
+    messages.push(getErrorMessages(this.operations?.entryCreate?.exceptions));
+    messages.push(getErrorMessages(this.operations?.setEdoc?.exceptions));
+    messages.push(getErrorMessages(this.operations?.setFields?.exceptions));
+    messages.push(getErrorMessages(this.operations?.setLinks?.exceptions));
+    messages.push(getErrorMessages(this.operations?.setTags?.exceptions));
+    messages.push(getErrorMessages(this.operations?.setTemplate?.exceptions));
+
+    return messages.filter(item => item).join(" ");
+  }
+}
+
+const OPERATION_ID_HEADER: string = "x-requestid";
+export class ProblemDetails extends generated.ProblemDetails {
+  extensions: any;
+}
+export class ApiException extends ApiExceptionCore  {
+  message: string;
+  status: number;
+  headers: { [key: string]: any; };
+  problemDetails: ProblemDetails;
+
+  constructor(message: string, status: number, response: string, headers: { [key: string]: any; }, result: any) {
+    super(message, status, headers, result);
+
+    let problemDetails = new generated.ProblemDetails();
+
+    if (result instanceof generated.CreateEntryResult) {
+      problemDetails = generated.ProblemDetails.fromJS({
+        "title": result.getSummary(),
+        "status": status,
+        "operationId": headers[OPERATION_ID_HEADER],
+      });
+
+      problemDetails.extensions = {
+        "createEntryResult": Object.assign({}, result)
+      }
+    } else { 
+      problemDetails = result != null && result.status !== undefined ? result : generated.ProblemDetails.fromJS({
+        "title": "HTTP status code " + status,
+        "status": status,
+        "operationId": headers[OPERATION_ID_HEADER],
+      });
+    }
+
+    this.message = problemDetails?.title ??  message;
+    this.status = status;
+    this.headers = headers;
+    this.problemDetails = problemDetails 
+  }
+
 }